Split admin registration and reorganise definitions
